fix : 하나의 유저가 여러개의 좌석에 대해 예매시, 각각 다른 예매 식별자로 저장되는 문제

+++ src/main/java/com/example/demo/dto/SeatStatusResponseDto.java
@@ -8,6 +8,6 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class SeatStatusResponseDto {
-    private Long screeningId;
+    private Integer screeningId;
     private List<SeatWithStatusDto> availableSeatWithPrice;
 }
